pyproject.toml
[project]
name = "langgraph-funcapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saim-Hassan786", email = "saimaiengineer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.70",
    "litellm>=1.60.6",
    "python-dotenv>=1.0.1",
]

[project.scripts]
langgraph-funcapi = "langgraph_funcapi:main"
agents = "langgraph_funcapi.main:flow"
agents1 = "langgraph_funcapi.main:flow_stream"
router = "langgraph_funcapi.routing:main"
parallel = "langgraph_funcapi.parallel:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
